Refactor payment components to remove createOrder prop; streamline order creation logic in PaymentSuccess component
